Added GuiCreateTeam and CheckBoxListLM, fixed GuiHelper PositionedTextData getter, LMJsonUtils not serializing text components with siblings properly

+++ src/main/lib/com/feed_the_beast/ftbl/lib/util/LMColorUtils.java
@@ -87,7 +87,7 @@ public class LMColorUtils
         }
     };
 
-    public static int getColorFromID(byte col)
+    public static int getColorFromID(int col)
     {
         return ID_COLORS[col & 0xFF];
     }
